requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
docopt==0.6.2
gTTS==2.3.2
idna==3.4
markdown2==2.4.6
pdfminer.six==20220524
Pillow==9.5.0
py-pdf-parser==0.10.2
pycparser==2.21
PyPDF2==3.0.1
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
Wand==0.6.10
